xinput: check if device_id is created

+++ 3rdparty/xinput_calibrator_pointercal.sh
@@ -27,24 +27,33 @@ if [ ! "x${detected_capes}" = "x" ] ; then
 		CALFILE="/etc/pointercal.xinput.${scan_ts}"
 		echo "xinput_calibrator: xinput_calibrator --list | grep ${scan_ts}"
 		device_id=`$BINARY --list | grep ${scan_ts} | sed 's/ /\n/g' | grep id | awk -F "id=" '{print $2}'`
-
-		echo "xinput_calibrator: using: xinput: id=${device_id}"
-
-		if [ -e $CALFILE ] ; then
-		  if grep replace $CALFILE ; then
-		    echo "Empty calibration file found, removing it"
-		    rm $CALFILE
-		  else
-		    echo "Using calibration data stored in $CALFILE"
-		    . $CALFILE && exit 0
-		  fi
+		if [ "x${device_id}" = "x" ] ; then
+			echo "xinput_calibrator: too fast... sleep 5..."
+			sleep 5
+			device_id=`$BINARY --list | grep ${scan_ts} | sed 's/ /\n/g' | grep id | awk -F "id=" '{print $2}'`
 		fi
 
-		echo "xinput_calibrator: xinput_calibrator --device ${device_id} --output-type xinput -v"
-		CALDATA=`$BINARY --device ${device_id} --output-type xinput -v | tee $LOGFILE | grep '    xinput set' | sed 's/^    //g; s/$/;/g'`
-		if [ ! -z "$CALDATA" ] ; then
-		  echo $CALDATA > $CALFILE
-		  echo "Calibration data stored in $CALFILE (log in $LOGFILE)"
+		if [ ! "x${device_id}" = "x" ] ; then
+			echo "xinput_calibrator: using: xinput: id=${device_id}"
+
+			if [ -e $CALFILE ] ; then
+			  if grep replace $CALFILE ; then
+			    echo "Empty calibration file found, removing it"
+			    rm $CALFILE
+			  else
+			    echo "Using calibration data stored in $CALFILE"
+			    . $CALFILE && exit 0
+			  fi
+			fi
+
+			echo "xinput_calibrator: xinput_calibrator --device ${device_id} --output-type xinput -v"
+			CALDATA=`$BINARY --device ${device_id} --output-type xinput -v | tee $LOGFILE | grep '    xinput set' | sed 's/^    //g; s/$/;/g'`
+			if [ ! -z "$CALDATA" ] ; then
+			  echo $CALDATA > $CALFILE
+			  echo "Calibration data stored in $CALFILE (log in $LOGFILE)"
+			fi
+		else
+			echo "xinput_calibrator: failed to get device_id..."
 		fi
 	fi
